pyproject.toml
[project]
name = "neoflow"
version = "0.0.21"
description = "NeoFlow is an AI-powered development assistant designed to streamline the software development process."
requires-python = ">=3.12"
dependencies = [
    "weaviate-client>=4.19.0",
    "weaviate-agents>=1.2.0",
    "ollama",
    "selenium>=4.39.0",
    "beautifulsoup4>=4.14.0",
    "requests>=2.32.0",
    "rich>=14.0.0",
    "pydantic>=2.12.0",
    "httpx>=0.28.0",
    "grpcio>=1.76.0",
    "pyyaml>=6.0",
    "prompt_toolkit>=3.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
]

[project.scripts]
neoflow = "neoflow.cli:main"

[tool.setuptools.packages.find]
include = ["neoflow", "neoflow.*"]

[tool.setuptools.package-data]
neoflow = ["py.typed"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B011"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'

[tool.coverage.run]
source = ["neoflow"]
omit = [
    "tests/*",
    "**/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
